* renamed CommunicationSimulatorsIT.java to SendAndReceiveIT and all corresponding files * moved apps to package org.eclips.test.app... * removed unnecessary entries in output_config.xml * reformatted a couple of files * minor adjustments mentioned in PR
