+++ test/mosaic-integration-tests/pom.xml
@@ -82,6 +82,7 @@
                         <id>create-test-jar-with-rules</id>
                         <configuration>
                             <includes>
+                                <include>org/eclipse/mosaic/test/app/**</include>
                                 <include>org/eclipse/mosaic/test/junit/**</include>
                             </includes>
                         </configuration>
