+++ test/mosaic-integration-tests/src/main/java/org/eclipse/mosaic/test/app/sendandreceive/SendCamAppCell.java
@@ -12,8 +12,7 @@
  *
  * Contact: mosaic@fokus.fraunhofer.de
  */
-
-package org.eclipse.mosaic.test.sendandreceive;
+package org.eclipse.mosaic.test.app.sendandreceive;
 
 import org.eclipse.mosaic.fed.application.ambassador.simulation.communication.CellModuleConfiguration;
 import org.eclipse.mosaic.lib.enums.DestinationType;
