+++ test/mosaic-integration-tests/src/main/java/org/eclipse/mosaic/test/app/sendandreceive/messages/SimpleV2xMessage.java
@@ -13,7 +13,7 @@
  * Contact: mosaic@fokus.fraunhofer.de
  */
 
-package org.eclipse.mosaic.test.sendandreceive.messages;
+package org.eclipse.mosaic.test.app.sendandreceive.messages;
 
 import org.eclipse.mosaic.lib.objects.v2x.EncodedPayload;
 import org.eclipse.mosaic.lib.objects.v2x.MessageRouting;
