--- test/scenarios/CommunicationSimulators/sumo/tiergarten.sumocfg
@@ -1,10 +1,0 @@
-<configuration>
-	<input>
-		<net-file value="tiergarten.net.xml" />
-		<route-files value="tiergarten.rou.xml" />
-	</input>
-	<time>
-		<begin value="0" />
-		<end value="10000" />
-	</time>
-</configuration>
