+++ test/scenarios/SendAndReceive/cell/cell_config.json
