+++ test/scenarios/SendAndReceive/sns/sns_config.json
