+++ test/scenarios/SendAndReceive/sumo/tiergarten.rou.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" ?>
+<routes>
+    <vType id="default" length="7.5" maxSpeed="50" minGap="2.0">
+        <carFollowing-Krauss accel="0.80" decel="4.5" sigma="0.5" tau="1"/>
+    </vType>
+    <vehicle id="1.0" depart="1.00">
+        <route id="0" edges="4068038_251150126_428788319_21487167 4068038_251150126_428788319_272365223 4068038_428788319_408194194_428788319 4068038_428788319_408194194_272256206 4068038_408194194_423839224_408194194 4068038_408194194_423839224_26738489 4068038_408194194_423839224_313006383 4068038_423839224_26704448_423839224 36337928_26704448_27537750_26704448 4609244_27537750_27537749_27537750 4609243_27537749_252864801_27537749 4609243_252864801_252864802_252864801 4609243_252864801_252864802_265786533 32935479_252864802_21487170_252864802 32935479_252864802_21487170_341364279 32935479_252864802_21487170_248919692 32935479_252864802_21487170_82654385 32935479_252864802_21487170_27011308 32935479_252864802_21487170_341364277 32935479_252864802_21487170_341364270 32935479_252864802_21487170_27011305"/>
+    </vehicle>
+</routes>
